Add Nawaya 2011 proceeding in US, immediately after he left Canada
Nawaya's probation in Canada began Dec 14, 2010, lasting 1 year

One month later, he was heard in United States Court of Appeals,
Ninth Circuit, reviewing two Immigration orders that had him
removed from the US. One of the orders appears to have originated
in the Department of Homeland Security. The Board of Immigration
Appeals held that it "lacked authority to apply equitable estoppel
against the Department of Homeland Security".

It looks like the student visa problem was one of the
reasons for the Immigration order, but its true origin was with
the DHS.

The submission date on Nawaya v Holder is Jan 14, 2011.